+++ src/console/pythonconsole.cs
@@ -16,7 +16,7 @@ namespace Python.Runtime
     /// </remarks>
     public sealed class PythonConsole
     {
-#if NET40
+#if NET461
         private static AssemblyLoader assemblyLoader = new AssemblyLoader();
 #endif
         private PythonConsole()
@@ -26,8 +26,8 @@ namespace Python.Runtime
         [STAThread]
         public static int Main(string[] args)
         {
-            // Only net40 is capable to safely inject python.runtime.dll into resources.
-#if NET40
+            // Only .NET Framework is capable to safely inject python.runtime.dll into resources.
+#if NET461
             // reference the static assemblyLoader to stop it being optimized away
             AssemblyLoader a = assemblyLoader;
 #endif
@@ -40,7 +40,7 @@ namespace Python.Runtime
             return i;
         }
 
-#if NET40
+#if NET461
         // Register a callback function to load embedded assemblies.
         // (Python.Runtime.dll is included as a resource)
         private sealed class AssemblyLoader